Display a u instead of the full upstream name.
But only when it's what we expect.  For instance, if we're on master and
upstream is origin/master.  If upstream was szak/master, we'd display
the full upstream ref.  The idea is to consume less space when we can,
leaving a little more room for the path.  But not at the loss of
clarity.