completion: make a couple of tweaks to _make_dir_complete
I tend to use a symlink for ~/projects, so that I can put my files on a
case-sensitive volume under Mac OS X.  However, my cdp alias wouldn't
complete for files under ~/projects because it was a link.  Instead, use
the logical path for the cd command, and use the realpath for completion
only.

Also, it turns out cdc has some completion assigned to it already, so
unset it in my config before defining cdc.