Add a hook to let gdbinit know that we're running on an Apple version.
Unfortunately, there's no way to tell at runtime whether you're running
on an Apple customized version of gdb or a stock version.  On Apple's,
the CPSR register for the ARM is treated as a structure, and it's not
for all other versions.  This allows me to inject a flag to detect
whether or not it's an Apple version of gdb.