shell-common: teach the prompt to look at the publishing branch
It turns out that the "upstream" branch is intended to point to the
eventual integration branch for your topic.  At the moment, Git doesn't
have a concept of @{publish}--but it's coming.  So, let's teach the
prompt to show the ahead/behind status relative to where we publish even
in the triangular workflow.  This makes more sense when working from
several different machines and when collaborating with others as it's a
good reminder of whether you need to update the remote branch, or bring
in any new changes.