Cope with libedit vs. readline issues in startup.py.
Not only does libedit use a different mechanism for binding, it also
uses a different history format.  To complicate matters more, Mac's
system python lies about whether it's using readline or libedit.  To be
precise, it seems Apple has chosen to build Python using the readline
compatibility library rather than exposing libedit outright.  This
complicates matters for detection purposes.  So we now look to see if
the readline module is in the system area on the Mac, and if so, treat
it libedit.  It's not a perfect heuristic because of the libedit's
readline compatibility bindings, but it's the best we've got.

Note: we can no longer switch on just the platform type because things
like Homebrew or MacPorts may use actual readline vs. the native
libedit.