zsh: ensure that the terminal is in application mode on the command line
It appears some applications can leave the terminal in a goofy state
where the key presses no longer correspond to their mapped values.  It
appears this is because the terminal is no longer in application
mode, so let's kick the terminal into application mode, if it isn't
already.  This appeared to happen most after a failed mock build.