user/jszakmeister: add a script to sync directories offline
This script will take a starting directory and an ending directory, and
determine which files to be be copied and removed to turn the starting
directory into the ending one.  It creates a tarball of this, and
generates a script that you can use to complete the transformation on
another machine elsewhere, provided you have a copy of the start
directory.