zsh: use history-beginning-search-backward/-forward
Apparently, up-/down-line-or-beginning-search don't work as advertised.
There's a call in there to move the cursor to the end of the line, thus
screwing up the ability to add more characters to the search.  This only
became apparent in newer zsh installations due to some configuration set
up by the distros.