shell-common: fix path shortening under newer bash shells
Somewhere between version 4.2.53 and 4.3.42 of bash, it started
expanding the tilde character in the parameter expansion I was using to
help shorten the current working directory name--negating the shortening
effect I was seeking.

Simply escaping the tilde with a backslash (\) doesn't work well for us,
since under zsh the backslash is kept and becomes visible in the prompt.
So we workaround the issue by making sure the tilde appears to be
surrounded by quotes.