shell-common: fix clang detection under OS X
It was accidentally triggering under Linux, and that wasn't the intent.
The intent was to put the pip and tox functions into place only if clang
was installed and we were running under OS X.  On newer Linux
distributions, command helpers were suggesting that you install gcc due
to the version check being done.  We now run it under `command` to help
prevent that sort of helpfulness.