shell-common: prevent command not found handlers while probing
The `command` command is not enough to prevent PackageKit's command not
found helper from running.  It turns out there's a
command_not_found_handler function that needs to be unset in order to
prevent the "helpful" suggestions while probing.  So this commit
introduces a _run_helper that does that and uses it when probing cc.