fix-bom-eol-exec: only show the modified files by default
Added a command-line argument to show all files being examined, if you
want to see the full list of files being considered.  Also added a dry
run option which actually does most of the work but doesn't write the
results, so you can get an idea of what would be modified.