git-clean-branches: be a bit more robust
We now avoid deleting the branch you are on when using `--auto`.  Also,
be more robust in handling the command line options.  Finally, if the
user provides a branch name, attempt to use it as the base to compare
against.  However, if must be an "approved" name to avoid accidentally
causing damage to the repository.