gitconfig: prevent pushes with submodules from being broken
If you're not careful, you can push the superproject before the
submodules and end up with the superproject in a broken state until the
submodules have been pushed.  Setting push.recurseSubmodules to 'check'
will let newer versions of Git check the submodules and make sure the
referenced commits are available.