zsh: work around the path_helper issues differently
Sourcing exports.sh forced re-setting up the `PATH`, which may have been
modified in `~/.etcenv`.  Instead, let's just save off the path and
restore it at the appropriate time.  This is really geared towards macOS
systems (and maybe other BSDs) that are using `path_helper` and
`zprofile` to set up the path, and we don't have the permissions to edit
the system-level `zprofile`.