shell-common: prevent errors with grep/egrep in bash
Bash does not like having a function and an alias, unless you put the
function keyword in front of the function.  Apparently, I knew this
before as I documented it at the top of the file.  Let's add a couple
more comments while we're at it, just as a reminder.