strong-tea: a new script to help keep the display on
The Windows performance is only okay... it prevents the screen from
locking but doesn't necessarily keep the screen awake.  More work to be
done there.

Under Linux, we attempt to find the right version of Python to run to
gain access to the dbus package.  We also try a couple of different
dbus interfaces, since older systems don't necessarily implement the
freedesktop ones.