update-common: cope with Git Bash and the MINGW environment better
`$HOME` may not evaluate to `Path.home()`, so we need to do a bit of
work to expand `~` ourselves in this environment.  In particular, pay
attention to the value of `$HOME` as `Path.expanduser()` ignores it for
the default Windows implementation.  In addition, we have to do a little
work to convert a MINGW path to a proper Windows one via the `cmd /c`
trick.